% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{estimate}
\alias{estimate}
\title{Estimate parameter.}
\usage{
estimate(object, pars, ...)
}
\arguments{
\item{object}{fitted model}

\item{pars}{names of parameters to estimate}

\item{...}{other parameters to pass}
}
\description{
Estimate parameter from fitted model.
}
