% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribs.R
\name{rbern}
\alias{rbern}
\title{Random generator for the Bernoulli distribution.}
\usage{
rbern(n, prob)
}
\arguments{
\item{n}{number of samples to generate}

\item{prob}{event probability}
}
\value{
an \code{integer} vector of 0 (non-event) and 1 (event)
}
\description{
Random generator for the Bernoulli distribution.
}
