% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribs.R
\name{rbvec}
\alias{rbvec}
\title{Generate a random binary vector.}
\usage{
rbvec(n, prob, e.min = 0)
}
\arguments{
\item{n}{size of binary vector}

\item{prob}{event probability
(not accounting for minimum event constraint)}

\item{e.min}{minimum number of events}
}
\value{
an \code{integer} vector of 0 and 1
}
\description{
Generate a random binary vector.
}
