% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribs.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Random generator for the Dirichlet distribution.}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{number of vectors to generate}

\item{alpha}{vector of parameters of the Dirichlet distribution}
}
\value{
a \code{matrix} in which each row vector is Dirichlet distributed
}
\description{
Random generator for the Dirichlet distribution.
}
