% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{rmvnorm_clusters}
\alias{rmvnorm_clusters}
\title{Generate random multivarate clusters}
\usage{
rmvnorm_clusters(ns, means)
}
\arguments{
\item{ns}{number of data points in each cluster}

\item{means}{centers of each cluster}
}
\value{
list containing matrix \code{X} and labels \code{cl}
}
\description{
Generate random multivarate clusters
}
\examples{
ns <- c(5, 8, 7)
means <- list(c(-6, 1), c(-1, -1), c(0, 4))
d <- rmvnorm_clusters(ns, means)
}
